#85 Updating test paths
