+++ api/tests/unit/config/fixtures/quizzes/html_quiz/QuizTest.php
@@ -1,0 +1,15 @@
+<?php
+namespace App\Tests\unit\config\fixtures\quizzes\html_quiz;
+
+use PHPUnit\Framework\TestCase;
+
+class QuizTest extends TestCase
+{
+    public function testDefault()
+    {
+        $quiz = require dirname(__DIR__)  .'/../../../../../config/fixtures/quizzes/html-quiz/quiz.php';
+
+        self::assertIsArray($quiz);
+    }
+
+}
